feat: 유저 아이디 수정 및 찾기, 코드 간단화.

+++ src/main/java/spring/springserver/domain/auth/service/AuthService.java
@@ -3,23 +3,25 @@ package spring.springserver.domain.auth.service;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import spring.springserver.domain.auth.data.request.ChangeUsernameRequest;
 import spring.springserver.domain.auth.data.request.GenerateTokenRequest;
-import spring.springserver.domain.auth.data.request.PasswordResetRequest;
 import spring.springserver.domain.auth.data.request.SignInRequest;
 import spring.springserver.domain.auth.data.request.SignUpRequest;
+import spring.springserver.domain.auth.data.request.FindUsernameRequest;
+import spring.springserver.domain.auth.data.request.PasswordResetRequest;
+import spring.springserver.domain.auth.data.response.FindUsernameResponse;
 import spring.springserver.domain.auth.data.response.PasswordResetResponse;
-import spring.springserver.domain.auth.data.response.SignOutResponse;
-import spring.springserver.domain.auth.data.response.SignUpResponse;
 import spring.springserver.domain.auth.data.response.SignInResponse;
+import spring.springserver.domain.auth.data.response.SignUpResponse;
+import spring.springserver.domain.auth.data.response.SignOutResponse;
+import spring.springserver.domain.auth.data.response.ChangeUsernameResponse;
 import spring.springserver.domain.auth.exception.AuthStatusCode;
 import spring.springserver.domain.member.entity.Member;
 import spring.springserver.domain.member.repository.MemberRepository;
 import spring.springserver.global.exception.exception.ApplicationException;
-import spring.springserver.global.jwt.JwtProvider;
 
 @Service
 @RequiredArgsConstructor
@@ -83,7 +85,7 @@ public class AuthService {
      * @return message
      */
     public SignOutResponse signOut(HttpServletRequest httpServletRequest,
-                                                 HttpServletResponse httpServletResponse) {
+                                   HttpServletResponse httpServletResponse) {
 
         tokenService.deleteTokens(
                 httpServletRequest,
@@ -93,40 +95,86 @@ public class AuthService {
         return SignOutResponse.of("로그아웃 되었습니다.");
     }
 
-    public PasswordResetResponse resetPasswordWithoutAuth(PasswordResetRequest request) {
+    public PasswordResetResponse resetPasswordWithoutAuth(PasswordResetRequest passwordResetRequest) {
 
-        Member member = memberRepository.findByUsername(request.username())
+        Member member = memberRepository.findByUsername(passwordResetRequest.username())
                 .orElseThrow(
                         () -> new ApplicationException(AuthStatusCode.USERNAME_NOT_FOUND)
                 );
 
-        String encoded = passwordEncoder.encode(request.newPassword());
-        member.setPassword(encoded);
+        member.setPassword(passwordEncoder.encode(passwordResetRequest.newPassword()));
 
         return PasswordResetResponse.of("비밀번호가 변경되었습니다.");
     }
 
     public PasswordResetResponse resetPasswordWithAuth(HttpServletRequest httpServletRequest,
                                                        HttpServletResponse httpServletResponse,
-                                                       PasswordResetRequest request) {
+                                                       PasswordResetRequest passwordResetRequest) {
+
+        String accessToken = tokenService.extractTokenFromCookie(
+                httpServletRequest,
+                "accessToken"
+        );
+
+        if (accessToken == null || accessToken.isBlank()) {
+            throw new ApplicationException(AuthStatusCode.INVALID_JWT);
+        }
+
+        Member member = memberRepository.findByUsername(passwordResetRequest.username())
+                .orElseThrow(
+                        () -> new ApplicationException(AuthStatusCode.USERNAME_NOT_FOUND)
+                );
+
+        String encoded = passwordEncoder.encode(passwordResetRequest.newPassword());
+        member.setPassword(encoded);
+
+        tokenService.deleteTokens(
+                httpServletRequest,
+                httpServletResponse
+        );
+
+        return PasswordResetResponse.of("비밀번호가 변경되었습니다. 다시 로그인 해주세요.");
+    }
+
+    public FindUsernameResponse findUsername(FindUsernameRequest findUsernameRequest) {
+
+        String username = memberRepository.findUsernameByEmail(findUsernameRequest.email())
+                .orElseThrow(
+                        () -> new ApplicationException(AuthStatusCode.USERNAME_NOT_FOUND)
+                );
+
+        return FindUsernameResponse.of(username);
+    }
+
+    public ChangeUsernameResponse resetUsernameWithAuth(HttpServletRequest httpServletRequest,
+                                                        HttpServletResponse httpServletResponse,
+                                                        ChangeUsernameRequest changeUsernameRequest) {
 
         String accessToken = tokenService.extractTokenFromCookie(httpServletRequest, "accessToken");
         if (accessToken == null || accessToken.isBlank()) {
             throw new ApplicationException(AuthStatusCode.INVALID_JWT);
         }
 
-        Member member = memberRepository.findByUsername(request.username())
+        Member member = memberRepository.findByUsername(tokenService.getCurrentUsername(httpServletRequest))
                 .orElseThrow(
                         () -> new ApplicationException(AuthStatusCode.USERNAME_NOT_FOUND)
                 );
 
-        String encoded = passwordEncoder.encode(request.newPassword());
-        member.setPassword(encoded);
+        if (!member.getEmail().equals(changeUsernameRequest.email())) {
+            throw new ApplicationException(AuthStatusCode.INVALID_CREDENTIALS);
+        }
+
+        if (memberRepository.existsByUsername(changeUsernameRequest.newUsername())) {
+            throw new ApplicationException(AuthStatusCode.USERNAME_ALREADY_EXIST);
+        }
+
+        member.setUsername(changeUsernameRequest.newUsername());
 
         tokenService.deleteTokens(
                 httpServletRequest,
-                httpServletResponse);
+                httpServletResponse
+        );
 
-        return PasswordResetResponse.of("비밀번호가 변경되었습니다. 다시 로그인 해주세요.");
+        return ChangeUsernameResponse.of("아이디가 변경 되었습니다. 다시 로그인 해주세요.");
     }
 }
